change file system structure to logically split dev-tools and app, create common folder for simple widely shared code
