+++ .npmignore
@@ -3,6 +3,5 @@ tsconfig.json
 .eslintrc.js
 .prettierrc.js
 .travis.yml
-example.ts
 jest.config.js
 .env
