+++ package.json
@@ -5,7 +5,7 @@
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "scripts": {
-    "dump": "ts-node src/utils/cli.ts",
+    "dump": "ts-node src/dev-tools/cli.ts",
     "scrape": "ts-node-dev src/scraper.ts",
     "test": "jest",
     "lint": "eslint src/**/*.ts",
