+++ src/common/assertions.ts
@@ -1,0 +1,11 @@
+import cheerio from 'cheerio';
+
+export const assertPageTitle = (expectedPageTitle: string, $: cheerio.Root) => {
+  const actualPageTitle: string = $('title')
+    .first()
+    .text();
+
+  if (expectedPageTitle != actualPageTitle) {
+    throw new Error(`wrong page type for parser\n\nExpected: ${expectedPageTitle}\nReceived: ${actualPageTitle}`);
+  }
+};
