+++ src/dev-tools/schedule-dump.ts
@@ -4,8 +4,8 @@ import { performance } from 'perf_hooks';
 
 import coursesJSON from '../../static/courses/courses.json';
 import { Course } from '../types';
-import { forEachHelper } from './common';
-import { classScheduleListingUrl } from '../lib/urls';
+import { forEachCourseHelper } from './helpers';
+import { classScheduleListingUrl } from './urls';
 
 export const scheduleUtil = async (term: string) => {
   const writeCourseScheduleToFS = async (course: Course) => {
@@ -18,18 +18,18 @@ export const scheduleUtil = async (term: string) => {
       // Not all courses in courses.json are found in the class schedule listing as courses.json
       // is retrieved from the Kuali API and course schedule listing from BAN1P.
       if (!fs.existsSync(destDir)) {
-          fs.mkdirSync(destDir, { recursive: true });
+        fs.mkdirSync(destDir, { recursive: true });
       }
 
       await fs.promises.writeFile(destFilePath, res.rawBody);
     }
-  }
+  };
 
   console.log('Starting schedule dump\n');
   const start = performance.now();
 
   const courses = coursesJSON as Course[];
-  await forEachHelper(courses, writeCourseScheduleToFS, 10);
+  await forEachCourseHelper(courses, writeCourseScheduleToFS, 10);
 
   const finish = performance.now();
   console.log(`Getting course data took ${(finish - start) / 60000} minutes`);
