+++ src/pages/courseListingTerm/__tests__/extractor.test.ts
@@ -1,10 +1,10 @@
+import appRoot from 'app-root-path';
 import * as cheerio from 'cheerio';
 import fs from 'fs';
 import path from 'path';
 
-import { extractCourseListingTerm } from '../pages/courseListingTerm/extractor';
-import { getSchedule } from '../utils/tests/getSchedule';
-import appRoot from 'app-root-path';
+import { extractCourseListingTerm } from '../extractor';
+import { getSchedule } from '../../../common/pathBuilders';
 
 // title: Course Listing Term
 // https://www.uvic.ca/BAN1P/bwckctlg.p_disp_cat_term_date
